Added Routing, Comments, Reactstrap, and Readme
installed NPM for reactstrap, react-router-dom. Which will allow for easier implementation of bootstrap, and will allow for React to change pages based on the URL.

Added to the readme.md file to help create easier implementation of this build for teammates.

Added commenting for JSX on files so we can create notes to help improve communcation May go horribly wrong becuase of versioning but no one seems interested in Meistertask.